Исправления согласно ревью
