--- service/src/main/java/ru/practicum/ewm/service/controller/admin/AdminController.java
@@ -1,140 +1,0 @@
-package ru.practicum.ewm.service.controller.admin;
-
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.ewm.service.dto.category.CategoryDto;
-import ru.practicum.ewm.service.dto.category.NewCategoryDto;
-import ru.practicum.ewm.service.dto.compilation.CompilationDto;
-import ru.practicum.ewm.service.dto.compilation.CompilationUpdateDto;
-import ru.practicum.ewm.service.dto.compilation.NewCompilationDto;
-import ru.practicum.ewm.service.dto.event.EventDto;
-import ru.practicum.ewm.service.dto.event.EventUpdateRequestDto;
-import ru.practicum.ewm.service.dto.user.NewUserDto;
-import ru.practicum.ewm.service.dto.user.UserDto;
-import ru.practicum.ewm.service.model.enums.EventState;
-import ru.practicum.ewm.service.service.interfaces.CategoryService;
-import ru.practicum.ewm.service.service.interfaces.CompilationService;
-import ru.practicum.ewm.service.service.interfaces.EventService;
-import ru.practicum.ewm.service.service.interfaces.UserService;
-
-import javax.validation.Valid;
-import javax.validation.constraints.Positive;
-import javax.validation.constraints.PositiveOrZero;
-import java.time.LocalDateTime;
-import java.util.List;
-
-@AllArgsConstructor
-@Slf4j
-@RestController
-@RequestMapping("/admin")
-public class AdminController {
-    private final UserService userService;
-    private final CategoryService categoryService;
-    private final CompilationService compilationService;
-    private final EventService eventService;
-
-    @PostMapping("/users")
-    @ResponseStatus(HttpStatus.CREATED)
-    public UserDto addUser(@Valid @RequestBody NewUserDto userDto) {
-        log.info("Создание пользователя: {}", userDto);
-        return userService.addUser(userDto);
-    }
-
-    @GetMapping("/users")
-    @ResponseStatus(HttpStatus.OK)
-    public List<UserDto> getAllUsers(
-            @RequestParam(required = false) List<Long> ids,
-            @RequestParam(name = "from", defaultValue = "0") @PositiveOrZero Integer from,
-            @RequestParam(name = "size", defaultValue = "10") @Positive Integer size
-    ) {
-        log.info("Вызов всех пользователей");
-        return userService.getAllUsers(ids, from, size);
-    }
-
-    @DeleteMapping("/users/{userId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable Long userId) {
-        log.info("Удаление пользователя: {}", userId);
-        userService.deleteUser(userId);
-    }
-
-    @PostMapping("/categories")
-    @ResponseStatus(HttpStatus.CREATED)
-    public CategoryDto addCategory(@Valid @RequestBody NewCategoryDto newCategoryDto) {
-        log.info("Создание категории: {}", newCategoryDto);
-        return categoryService.addCategory(newCategoryDto);
-    }
-
-    @PatchMapping("/categories/{catId}")
-    @ResponseStatus(HttpStatus.OK)
-    public CategoryDto updateCategory(
-            @PathVariable
-            Long catId,
-            @Valid @RequestBody
-            CategoryDto categoryDto
-    ) {
-        log.info("Обновление категории: {}, to: {}", catId, categoryDto);
-        return categoryService.updateCategory(catId, categoryDto);
-    }
-
-    @DeleteMapping("/categories/{catId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCategory(@PathVariable Long catId) {
-        log.info("Удаление категории: {}", catId);
-        categoryService.deleteCategory(catId);
-    }
-
-    @PostMapping("/compilations")
-    @ResponseStatus(HttpStatus.CREATED)
-    public CompilationDto addCompilation(@Valid @RequestBody NewCompilationDto newCompilationDto) {
-        log.info("Создание компиляции: {}", newCompilationDto);
-        return compilationService.addCompilation(newCompilationDto);
-    }
-
-    @PatchMapping("/compilations/{compId}")
-    @ResponseStatus(HttpStatus.OK)
-    public CompilationDto updateCompilation(
-            @PathVariable Long compId,
-            @Valid @RequestBody CompilationUpdateDto updateCompilationDto
-    ) {
-        log.info("Обновление компиляции с: {}, на: {}", compId, updateCompilationDto);
-        return compilationService.updateCompilation(compId, updateCompilationDto);
-    }
-
-    @DeleteMapping("/compilations/{compId}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCompilation(@PathVariable Long compId) {
-        log.info("Удаление компиляции: {}", compId);
-        compilationService.deleteCompilation(compId);
-    }
-
-    @PatchMapping("/events/{eventId}")
-    @ResponseStatus(HttpStatus.OK)
-    public EventDto updateEvent(
-            @PathVariable Long eventId,
-            @Valid @RequestBody EventUpdateRequestDto requestDto
-    ) {
-        log.info("Обновление события с: {}, на: {}", eventId, requestDto);
-        return eventService.updateEventByAdmin(eventId, requestDto);
-    }
-
-    @GetMapping("/events")
-    @ResponseStatus(HttpStatus.OK)
-    public List<EventDto> getAllEvents(
-            @RequestParam(required = false) List<Long> users,
-            @RequestParam(required = false) List<Long> categories,
-            @RequestParam(required = false) List<EventState> states,
-            @RequestParam(required = false)
-            @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeStart,
-            @RequestParam(required = false)
-            @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss") LocalDateTime rangeEnd,
-            @RequestParam(name = "from", defaultValue = "0") @PositiveOrZero Integer from,
-            @RequestParam(name = "size", defaultValue = "10") @Positive Integer size
-    ) {
-        log.info("Вызов всех событий");
-        return eventService.getAllEventsByAdmin(users, categories, states, rangeStart, rangeEnd, from, size);
-    }
-}
